JENKINS-36919: only count containers provisioned by Hudson
Other Docker clients may create containers on the docker host of our
Jenkins using the same image as Jenkins will do. Before, these
containers were considered when counting both for the Instance Capacity
and the Container Cap. Especially in SWARM-based situation where
multiple Jenkins instances are using the same docker host, this may lead
to unwanted side-effects.
The commit introduces a synchronized set, which "remembers" the
containerIds, which were provisioned by this Jenkins instance. When
determining the number of running containers (Container Cap) and/or the
number of containers of the selected image (Instance Capacity), the set
is used to only count those containers, which originate from this
Jenkins instance.