[FEATURE] Introduce "enabled" checkbox on Docker Cloud configuration
The "container cap" configuration option has a changeful history.
Initially, there was a distinction between setting it to zero (do not
provision any container) and to leave it blank (limit is disabled). In
recent times, there is no such distinction happening anymore (blank
means zero and means "do not provision anything"), but the documentation
indicated that this was the way to allow infinite provisioning. 
With the previous commit, the behaviour has been changed to accept
provisioning without limit, if the limit was set to zero. However, this
removed the option to indicate that no provisioning shall take place
(may happen in some scenarios, e.g. if you want to cease creating
containers temporarily, for instance due to maintenance).

To make it more transparent, a new checkbox "Enabled" is introduced to
the docker cloud. By default, it is selected and thus allows that new
containers are provisioned. If the selection is removed from the
checkbox, no provioning of new containers will take place (already
runnign containers are still handled as before).